requirements.txt
boto3==1.34.32
Pillow==8.3.1
